[TM-32] Document Architecture and Pipeline
Problem: document the packages that Morley depends on.

Solution: create docs/dependencies.md which lists the package name, the
package link in Hackage, an edited snippet of what it does from the
package description, motivation (if not clear) for using this package
(missing this, need to update), and any useful tutorials for using the
package.