[TM-58] Add special handling of PAPAIR
Problem: when we parse PAIR we treat it as a PAPAIR macro.
It's an old hack. Because of that when we parse a contract with PAIR
we get a different contract: PAIR turns into { PAIR } (a sequence with
one instruction).
Solution: as an easy solution we handle PAIR specially.
Later we should come up with something better.