[TM-35] Improve "odd lengthed bytes" error
Problem: on attempt to parse `0x000` a confusing error is raised.

This happens because, when we fail with `OddNumberBytesException`, that
happens within `try` block and thus we backtrack and try to parse further.

Solution: fail outside `try`, since if we parsed some `0x.*`, we have
to return or fail.
Also made error message look like in `./alphanet.sh`.