[TM-35] Increase expected gas consumption in environment test
As in the reference implementation, all instruction blocks are now
always wrapped into `Nested`, and this adds extra step at
interpretation.

Just recently it has been decreased to 19, so I think it's fine to
return it back to 20.

I also inserted some `HasCallStack`s, since when tests fail they
display bottom of the known call stack, and knowing exact place
where the test is initiated simplifies search for a developer.