[TM-126] Use tasty more actively
Problem: in the previous commit we have changed our primary testing
engine to `tasty`, but we are still using `hspec` internally which
serves the same purpose as `tasty`. Ideally we shouldn't use both.

Solution: this commit eliminates all usages of `hspec` in places where
it's easy to do. Note that we are still using `hspec-expectations` in
some places, because it's just a small wrapper over HUnit with some
convenient helpers for testing, not a full-featured testing framework
like `hspec` or `tasty`.

Now in some places we don't need to group all tests manually, it's
done by `tasty-discover`. We still need to explicitly write them in
export lists. It ensures that if some test is not discovered (e. g.
due to a typo in its name or due to some changes in `tasty-discover`
convention), we'll get a warning. Either from GHC (unused top-level
binding) or from weeder (unused export). If we permit implicit
exports in testing modules, some test can be silently missed.
Another option is to always export only one thing (e. g. `Spec` or
`TestTree`), but I find it worse, because you need to add some name
to such group of tests, which most likely will be redundant, because
module names are already used to group and name tests. Also this way
we can't accidentally run a test more than once.

One unpleasant thing that I had to do is to add `Test.Tasty.HUnit`
which is a subset of the `tasty-hunit` package. The reason is that
`tasty-hunit` doesn't know about `HUnit` and because of that it
doesn't interoperate with tests which use `HUnit` directly or any
library that uses `HUnit` (e. g. `hspec-expectations`). Specifically,
it defines its own `HUnitFailure` and ignores HUnit's `HUnitFailure`.
I find it desirable to still be able to write `HUnit` tests and
use `hspec-expectations`, so I decided not to use `tasty-hunit` and
define `Test.Tasty.HUnit`, which is quite small and simple.
Note that we not only use it directly, but also it's used by
`tasty-discover`.