[TM-126] Add --fail-fast option to our test suite
Problem: sometimes it's convenient to be able to stop test suite as
soon as first failure is encountered. But we don't have such option.
Solution: use `tasty-fail-fast` which defines `failFast` decorator.
We can't use autogenerated `main` because `failFast` is not
`Ingredient` but a function and because generated `main` can only
append custom ingredients to default ones, not to modify them.
So now we use `tasty-discover` only to generate `tests :: IO TestTree`
and define `main` manually.