[TM-121] Implement registry service
This commit implements registry service as it is done in `dstoken` repo.

Services it depends on are not implemented yet, so I placed stubs.
Trust service will be implemented on top of the registry service in
TM-122.

Tests are to be implemented in TM-158 and TM-159.

While implementing registry service I followed the following sources:
* [Interface](https://gitlab.com/morley-framework/dstoken/blob/20f842bb/contracts/legacy/registry/DSRegistryServiceInterface.sol)
* [Implementation](https://gitlab.com/morley-framework/dstoken/blob/20f842bb/contracts/legacy/registry/ESRegistryService.sol)
* [Tests](https://gitlab.com/morley-framework/dstoken/blob/20f842bb/test/ESRegistryService.test.js)
(it may be beneficial to read all of these in order to understand better
what is going on).