[TM-160] Discourage usage of 'fail_'
Problem: 'fail_' (analog of FAIL) macro is bad because it always
fails with unit and doesn't carry any information about failure.
We'd like to avoid its usage.
Solution: write a comment discouraging its usage and raise a warning
when it's used. Also get rid of its usages.
One can still use it locally as the quickest way to fail.