[TM-120] Add new macros
Problem: if we have a function which wants to work with part of our
current stack, it's better to put this part to bottom and let this
function know the whole stack. Otherwise it won't be able to use
the functions from Lorentz.Referenced. Currently it's not very
convenient to move items to the top of the stack.
Solution: add 'elevateX' macro to move an item to the top of the stack
and 'dropX' which is used by 'elevateX'.

We also need a function to move an item back to the bottom of the
stack, but I feel that I've already spent too much time, so it will
be done in TM-182.