[TM-214] Develop PoC of high level language over Lorentz
Problem: It's kinda uncomfortable to develop in Lorentz
because you have to manipulate by stack directly.

Solution: Develop PoC of a high level language over Lorentz
where possible to create a variables.
For now, only allocating and setting of variables is possible.
Also, 'if' and 'while' statements are implemented to show that
developed approach is comprehensive.
Also, 'Expr' datatype is provided which makes possible to construct expressions
and compile them to Lorentz.
Developed language is called Indigo.