[TM-230] Store events in the storage
Problem: DS protocol requires that it should be possible to obtain list
of all events that happened to the contract. For example, addition of an
investor or issuance of tokens.

Solution: Store events in the contract storage. Add entry point to
`DSProtocol` itselft, which is used for getting all the events,
that happened.