[TM-276] Add module to work with `Columnar`
Problem: for selected way to migrate storage we will need to
be able to use UStore template both for initialization and for
migration. Columnars fit good here. Moreover, we will need to
traverse all fields of storage and collect migrations for those fields.

Also, we already have some logic with columnars in DS Token.

Solution: put generalized logic of working with columnars in an utility
module.