[TM-276] Simplify UStore initializer
Now when we use columnars in `UStore`, we can abandon `UStoreSubMap`
and `UStoreField` constructors and provide only required values
themselves in UStore initializer.

However, this makes the code which provides such initialization slightly
more difficult because we can't just traverse the template and
understand the meaning of each field (submap or just field),
fields can have arbitrary types now.
For this sake in generic traversal we simultaneously traverse the
template with 'Columnar f' part cut off and the initializer for
'UStore'.