[TM-64] Print annotations from typed contracts
Problem: Typed `Instr` does not have fields to store annotations

Solution: This change just extract annotations for the storage and
parameter from a typed contract's (SomeContract) input stack type
(HST), and add it to the untyped contract's storage and parameter
fields. This does not actually add annotations to the instructions or
types, but just include the annotations for storage and parameter fields
while printing the typed contract.

Tests: A test was added that read a contract with some annotations for
contract parameters, converted it to typed version, and printed it and
checked that the output contains the annotations.