[TM-64] Add annotations to typed contracts
Problem: Typed `Instr` does not have fields to store annotations

Solution: A new constructor was added to the `Michelson.Typed.T.T` type
`TFAnnotated TFieldAnn T`, which would wrap types that occupy annotated
fields.

Another approach, which added fields directly to the `Option`,
`Pair` and `Or` constructors was tried, but it turned out to result in
a lot more changes than the current approach. The code that converted
untyped types to `T` was changes to wrap the resulting type in
`TFAnnotated` constructor, if the field annotations are not empty.

The `TFieldAnn` type is a sum type that wraps a text in one
branch and a `Symbol` in other, so that it can represent the string
annotation both in type and term levels.

Since the promoted constructors were used in type checking of the
instruction, it was required to discard the annotation wrapped types
during type checking, and only consider their wrapped types. For this
the `eqType' function was modified to compare the types, after erasing
the annotations using a type family. Then the resulting witness was
`unsafeCoerce`ed into a witness for the equivalence original, unwrapped
types. This appeared to be safe thing to do since the `eqType` function
was restricted to work with types of kind `T`, and how the `EraseNote`
typefamily is implemented. Another approach that was tried was to return
and equivalence for `EraseNote a` and `EraseNote b` instead of `a` and
`b`, and have the rest of the code work with this equivalence. But this
turned out to be too extensive of a change, not to mention the
overwhelming amount of type errors that just kept on popping up.

Tests: A test was added that read a contract with some annotations for
contract parameters, converted it to typed version, and printed it and
checked that the output contains the annotations.