[TM-64] Retain notes in typed contract
Problem: Typed `Instr` does not have fields to store annotations

Solution: This change just extract typed annotations returned by each
iteration of the instruction type checker and wraps it along with the
typed instruction. To do this a new constructor shown below was added
to the typed instruction type.

```
  data PackedNotes a =
    PackedNotes
      { pnNotes :: Notes a
      , pnSing :: Maybe (Sing a)
      }

  SeqWithNotes
    :: PackedNotes (StackHead a)
    -> Instr a b
    -> Instr b c
    -> Instr a c
```

This means that the typed contract is now a tree of `SeqWithNotes`
instructions instead of `Seq` instructions. And at each such node, the
`SeqWithNotes` will contain the annotations for the type that is the
input of the instruction it wraps.

In this way, when the final type checked contract is availble, we will
have the annotations for the contract input at the very top node of the
tree.

Tests: A test was added that read a contract with some annotations for
contract parameters, converted it to typed version, and printed it and
checked that the output contains the annotations for contract
parameters.